refactor: rename api data
