+++ RyoURL/shortURL/apis/__init__.py
@@ -1,2 +1,2 @@
-from .auth import auth_router
-from .short import url_router
+from .auth_api import auth_router
+from .short_api import url_router
